Fix the websocket!
Yes it looks weird, let me explain.

First, in the constructor, we make sure we capture ALL Websocket send events (no worries, there are only 2 websockets). Then we add it to an array of sockets if we haven't seen it before. if it is the spotify WebSocket, we trigger the webSocketPatch function. This will add a listener to the WebSocket and remove the function that captures all WebSocket Send events. On stop, both the WebSocket Send capture function get removed and the eventListener gets removed.